CLI for migration, pid_swapping
